+++ prototypes/djerba_modular/plugin_wise/__init__.py
@@ -1,3 +1,0 @@
-"""Define a name for the plugin"""
-
-PLUGIN_NAME = 'plugin_wise'