+++ src/lib/djerba/plugins/case_overview/test/plugin_test.py
@@ -1,9 +1,10 @@
 #! /usr/bin/env python3
 
-"""Test of the patient info plugin"""
+"""Test of the case overview plugin"""
 
-import logging
 import os
+import shutil
+import logging
 import string
 import unittest
 import tempfile
@@ -19,28 +20,30 @@ class TestCaseOverview(PluginTester):
         self.maxDiff = None
         self.tmp = tempfile.TemporaryDirectory(prefix='djerba_')
         self.tmp_dir = self.tmp.name
-        sup_dir_var = 'DJERBA_TEST_DATA'
+        sup_dir_var = 'DJERBA_TEST_DIR'
         self.sup_dir = os.environ.get(sup_dir_var)
+        self.sample_json = os.path.join(self.sup_dir, "plugins/case_overview/sample_info.json")
 
     def testCaseOverviewWGTS(self):
         test_source_dir = os.path.realpath(os.path.dirname(__file__))
-        json_location = os.path.join(self.sup_dir ,"case-overview-plugin/case_overview_WGTS.json")
+        json_location = os.path.join(self.sup_dir ,"plugins/case_overview/case_overview_WGTS.json")
+        shutil.copy(self.sample_json, self.tmp_dir)
 
         params = {
             self.INI: 'case_overview_WGTS.ini',
             self.JSON: json_location,
-            self.MD5: 'f8763747e45463c8163a8fe6d6c1956a' # TO CHANGE
+            self.MD5: '7532366678d57cc4ac51ac790cc3f902'
         }
         self.run_basic_test(test_source_dir, params)
 
     def testCaseOverviewTAR(self):
         test_source_dir = os.path.realpath(os.path.dirname(__file__))
-        json_location = os.path.join(self.sup_dir ,"case-overview-plugin/case_overview_TAR.json")
+        json_location = os.path.join(self.sup_dir ,"plugins/case_overview/case_overview_TAR.json")
 
         params = {
             self.INI: 'case_overview_TAR.ini',
             self.JSON: json_location,
-            self.MD5: 'f8763747e45463c8163a8fe6d6c1956a' # TO CHANGE
+            self.MD5: '08627c99b146c91c034f56c90563570b'
         }
         self.run_basic_test(test_source_dir, params)