+++ src/lib/djerba/plugins/cnv/tools.py
@@ -110,7 +110,8 @@ class cnv_processor(logger):
         image_converter = converter(self.log_level, self.log_path)
         cnv_plot = image_converter.convert_svg(self.plot_path, 'CNV plot')
         rows = []
-        is_wgts = wgts_tools.has_expression(self.work_dir)
+        wgts_toolkit = wgts_tools(self.log_level, self.log_path)
+        is_wgts = wgts_toolkit.has_expression(self.work_dir)
         if is_wgts:
             self.logger.info("Reading expression from {0}".format(self.work_dir))
             mutation_expression = wgts_tools.read_expression(self.work_dir)
@@ -123,18 +124,18 @@ class cnv_processor(logger):
             reader = csv.DictReader(input_file, delimiter="\t")
             for row_input in reader:
                 gene = row_input[self.HUGO_SYMBOL_UPPER_CASE]
+                # if gene not found in cytoBands.txt, default to 'Unknown'
                 row_output = {
                     cnv.EXPRESSION_PERCENTILE: mutation_expression.get(gene), # None for WGS
                     wgts_tools.GENE: gene,
                     cnv.GENE_URL: html_builder.build_gene_url(gene),
                     cnv.ALTERATION: row_input[self.ALTERATION_UPPER_CASE],
-                    wgts_tools.CHROMOSOME: cytobands.get(gene),
+                    wgts_tools.CHROMOSOME: cytobands.get(gene, wgts_tools.UNKNOWN),
                     wgts_tools.ONCOKB: oncokb_levels.parse_oncokb_level(row_input)
                 }
                 rows.append(row_output)
         unfiltered_cnv_total = len(rows)
         self.logger.debug("Sorting and filtering CNV rows")
-        wgts_toolkit = wgts_tools(self.log_level, self.log_path)
         rows = list(filter(oncokb_levels.oncokb_filter, wgts_toolkit.sort_variant_rows(rows)))
         results = {
             cnv.PERCENT_GENOME_ALTERED: self.calculate_percent_genome_altered(),