+++ src/lib/djerba/plugins/demo1/test/plugin_test.py
@@ -13,7 +13,7 @@ class TestDemo1(PluginTester):
         params = {
             self.INI: 'demo_1.ini',
             self.JSON: 'demo_1.json',
-            self.MD5: 'c313b41172656e9b6d377ffbd6fe0332'
+            self.MD5: '9aaa43ddde3e4ac53b9d15b7eb43df7a'
         }
         self.run_basic_test(test_source_dir, params)