+++ src/lib/djerba/plugins/failed_report/test/plugin_test.py
@@ -23,7 +23,7 @@ class TestFailedReportPlugin(PluginTester):
         params = {
             self.INI: 'failed_report.ini',
             self.JSON: json_location,
-            self.MD5: '80cc6991ac78ba40884b30b5436346f5'
+            self.MD5: 'fb99e52942e6ef66c710d848c45a7f48'
         }
         self.run_basic_test(test_source_dir, params)