+++ src/lib/djerba/plugins/supplement/body/test/plugin_test.py
@@ -9,62 +9,82 @@ import tempfile
 from djerba.util.validator import path_validator
 from djerba.plugins.plugin_tester import PluginTester
 
-class TestPwgSupplementaryPlugin(PluginTester):
+class TestSupplementaryPluginBody(PluginTester):
     def setUp(self):
         self.path_validator = path_validator()
         self.maxDiff = None
         self.tmp = tempfile.TemporaryDirectory(prefix='djerba_')
         self.tmp_dir = self.tmp.name
-        sup_dir_var = 'DJERBA_TEST_DATA'
+        sup_dir_var = 'DJERBA_TEST_DIR'
         self.sup_dir = os.environ.get(sup_dir_var)
 
     def testPwgsSupplementary(self):
         test_source_dir = os.path.realpath(os.path.dirname(__file__))
-        json_location = os.path.join(self.sup_dir ,"supplement-plugin/report_json/pwgs.supplement.json")
+        json_location = os.path.join(self.sup_dir ,"plugins/supplement/report_json/pwgs.supplement.json")
         params = {
             self.INI: 'PWGS.supp.ini',
             self.JSON: json_location,
-            self.MD5: 'd15b300a8034e1c8736e4dc9f300149e'
+            self.MD5: '3d74e80e39b34d073ce8be1e3b7ff8ea'
         }
         self.run_basic_test(test_source_dir, params)
 
     def testTarSupplementary(self):
         test_source_dir = os.path.realpath(os.path.dirname(__file__))
-        json_location = os.path.join(self.sup_dir ,"supplement-plugin/report_json/tar.supplement.json")
+        json_location = os.path.join(self.sup_dir ,"plugins/supplement/report_json/tar.supplement.json")
         params = {
             self.INI: 'TAR.supp.ini',
             self.JSON: json_location,
-            self.MD5: 'bf0f88cd88bb3a12dc9bf5651cd8547c'
+            self.MD5: '3eb160855c2f799967ee701679b29768'
         }
         self.run_basic_test(test_source_dir, params)
    
     def testTarFailSupplementary(self):
         test_source_dir = os.path.realpath(os.path.dirname(__file__))
-        json_location = os.path.join(self.sup_dir ,"supplement-plugin/report_json/tar.fail.supplement.json")
+        json_location = os.path.join(self.sup_dir ,"plugins/supplement/report_json/tar.fail.supplement.json")
         params = {
             self.INI: 'TAR.FAIL.supp.ini',
             self.JSON: json_location,
-            self.MD5: 'b56abd612a4c74bdf114ec3315436940'
+            self.MD5: '6a29ab5c95933f60a33c2ca3f5da6987'
         }
         self.run_basic_test(test_source_dir, params)
 
     def testWgtsSupplementary(self):
         test_source_dir = os.path.realpath(os.path.dirname(__file__))
-        json_location = os.path.join(self.sup_dir ,"supplement-plugin/report_json/wgts.supplement.json")
+        json_location = os.path.join(self.sup_dir ,"plugins/supplement/report_json/wgts.supplement.json")
         params = {
             self.INI: 'WGTS.supp.ini',
             self.JSON: json_location,
-            self.MD5: 'd3c2ec65780bfea9be4e9e27e8d67958'
+            self.MD5: 'fb65da5ca81b71d632bec84a699f6715'
         }
         self.run_basic_test(test_source_dir, params)
 
     def testWgtsFailSupplementary(self):
         test_source_dir = os.path.realpath(os.path.dirname(__file__))
-        json_location = os.path.join(self.sup_dir ,"supplement-plugin/report_json/wgts.fail.supplement.json")
+        json_location = os.path.join(self.sup_dir ,"plugins/supplement/report_json/wgts.fail.supplement.json")
         params = {
             self.INI: 'WGTS.FAIL.supp.ini',
             self.JSON: json_location,
-            self.MD5: 'dda3cab0258e6d09fd17d59706652740'
+            self.MD5: '56cd6255b9f68d4338423e735359c76e'
+        }
+        self.run_basic_test(test_source_dir, params)
+
+    def testWgts40XSupplementary(self):
+        test_source_dir = os.path.realpath(os.path.dirname(__file__))
+        json_location = os.path.join(self.sup_dir ,"plugins/supplement/report_json/wgts40X.supplement.json")
+        params = {
+            self.INI: 'WGTS40X.supp.ini',
+            self.JSON: json_location,
+            self.MD5: '9e1e21e55b4e8e5618c7a88ec468344c'
+        }
+        self.run_basic_test(test_source_dir, params)
+
+    def testWgts40XFailSupplementary(self):
+        test_source_dir = os.path.realpath(os.path.dirname(__file__))
+        json_location = os.path.join(self.sup_dir ,"plugins/supplement/report_json/wgts40X.fail.supplement.json")
+        params = {
+            self.INI: 'WGTS40X.FAIL.supp.ini',
+            self.JSON: json_location,
+            self.MD5: 'c514d5fc0313a75411315ec91c07c23b'
         }
         self.run_basic_test(test_source_dir, params)