+++ src/test/core/test_core.py
@@ -30,7 +30,7 @@ class TestCore(TestBase):
 
     LOREM_FILENAME = 'lorem.txt'
     SIMPLE_REPORT_JSON = 'simple_report_expected.json'
-    SIMPLE_REPORT_MD5 = '904bffdedff29e9ca16872d45ed12d21'
+    SIMPLE_REPORT_MD5 = 'f78e83206c5c12b989e1cee69503ffce'
     SIMPLE_CONFIG_MD5 = 'ab4b71b790f2b12aa802b8eaa1658951'
 
     class mock_args:
@@ -62,6 +62,7 @@ class TestCore(TestBase):
         with open(json_path) as json_file:
             data_found = json.loads(json_file.read())
             data_found['core']['extract_time'] = 'placeholder'
+            data_found['core']['core_version'] = 'placeholder'
         self.assertEqual(data_expected, data_found)
 
     def assertSimpleReport(self, json_path, html_path):
@@ -591,6 +592,7 @@ class TestSimpleReport(TestCore):
         config = djerba_main.configure(ini_path)
         data_found = djerba_main.extract(config)
         data_found['core']['extract_time'] = 'placeholder'
+        data_found['core']['core_version'] = 'placeholder'
         with open(json_path) as json_file:
             data_expected = json.loads(json_file.read())
         self.assertEqual(data_expected, data_found)